requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
inflection==0.3.1
iniconfig==2.0.0
lazy-object-proxy==1.10.0
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.7.0
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytest-mock==3.12.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
wrapt==1.16.0
